Example Gaussian Process prior parameter table

Example of the parameter table used for to fit the Gaussian Process prior model with virosolver. This data frame is used to control everything related to the model parameters, including values, names, fixed/estimates, and uniform prior bounds. For this parameter table, we have entries for the GP prior parameters (the latent states for each day "prob", the hyperparameters for the exponential kernel function, rho and nu), as well as entries for the viral kinetics/Ct model (peak Ct value, waning duration, time to peak, variation about the model Ct curve etc).

data(example_gp_partab)

Format

A data frame with 384 rows and 8 variables:

values

numeric values of the parameters

names

string names of the model parameters

fixed

binary values indicating if the parameter should be fixed (1) or estimated (0) during the MCMC procedure

lower_bound

lower numeric bound for the parameter during fitting (lower uniform prior bound)

upper_bound

upper numeric bound for the parameter during fitting (upper uniform prior bound)

steps

value between 0 and 1, giving the initial step size in the MCMC proposals. Note that these are adapted automatically

lower_start

can be used to set the lower allowable random starting value for the MCMC

upper_start

can be used to set the upper allowable random starting value for the MCMC

See also