Subset line list data by testing strategy. Options:

Sample a random fraction of the population if the only argument is frac_report

Sample some random fraction of the population at a subset of time points, specified by timevarying_prob

Observe symptomatic individuals with some fixed probability, frac_report if symptomatic is TRUE

Observe symptomatic individuals with some time-varying probability, timevarying_prob, if symptomatic is TRUE INPUTS:

individuals: the full line list from the simulation, returned by virosolver::simulate_observations_wrapper

solve_times: vector of times at which individuals can be reported

frac_report: the overall fraction/probability of individuals who are reported

timevarying_prob: a tibble with variables t and prob. This gives the probability of being reported on day t

symptomatic: if TRUE, then individuals are reported after developing symptoms. If FALSE, then we take a random cross-section OUTPUTS:

A tibble with line list data for individuals who were observed

A plot of incidence for both observed individuals and the entire simulated population

Plot growth rate of cases/infections in the entire population and observed population

Subset line list data by testing strategy. Options:

Sample a random fraction of the population if the only argument is frac_report

Sample some random fraction of the population at a subset of time points, specified by timevarying_prob

Observe symptomatic individuals with some fixed probability, frac_report if symptomatic is TRUE

Observe symptomatic individuals with some time-varying probability, timevarying_prob, if symptomatic is TRUE INPUTS:

individuals: the full line list from the simulation, returned by virosolver::simulate_observations_wrapper

solve_times: vector of times at which individuals can be reported

frac_report: the overall fraction/probability of individuals who are reported

timevarying_prob: a tibble with variables t and prob. This gives the probability of being reported on day t

symptomatic: if TRUE, then individuals are reported after developing symptoms. If FALSE, then we take a random cross-section OUTPUTS:

A tibble with line list data for individuals who were observed

A plot of incidence for both observed individuals and the entire simulated population

Plot growth rate of cases/infections in the entire population and observed population

simulate_reporting(
  individuals,
  solve_times,
  frac_report = 1,
  timevarying_prob = NULL,
  symptomatic = FALSE
)